Move resource screen functions into a trait
